feat: update eslint
