+++ .eslintrc.json
@@ -38,6 +38,15 @@
     "@typescript-eslint/no-implicit-any-catch": "error",
     "@typescript-eslint/no-dynamic-delete": "error",
     "@typescript-eslint/no-explicit-any": "error",
+    "max-lines": ["error", { "max": 100, "skipBlankLines": true, "skipComments": true }],
+    "max-depth": ["error", 3],
+    "max-lines-per-function": ["error", { "max": 70, "skipBlankLines": true, "skipComments": true }],
+    "max-params": ["error", { "max": 4 }],
+    "no-alert": "error",
+    "no-console": "error",
+    "no-delete-var": "error",
+    "no-const-assign": "error",
+    "no-unreachable": "error",
     "no-magic-numbers": ["error", { "ignore": [-1, 0, 1] }],
     "@typescript-eslint/no-redeclare": "error",
     "@typescript-eslint/naming-convention": [
